clash: avoid subshells for attr defs when possible
Each method call now checks if it is called from within another call by
checking an internal variable. If set we consider that all the local
variables needed by the object instance can be inherited and thus we can
skip calling their getters. That means that changing the value of the
local attribute variable in a method call will affect any next method call of
the same object instance, this is something to keep in mind.

Constructor attribute values are also known when creating the object, we
can always skip calling the getters and declare them with their known
values.

Performance gains:

When creating 1'000 elements in a list, this allows bash to jump from 16
seconds to around 11 seconds, with less time spent in the kernel, the number
of clone syscalls going down from 18'000 to 10'000 (roughly the same for
all shells except ksh which never forks for subshells that do not need
it).

This is still crazy slow, but better than before.

Going to 10'000 elements shows some interesting results, notably ksh
struggling to scale despite being close to 0 forks, jumping to almost
2G of virtual memory used, and taking the same amount of time than bash
(around 1min40), despite ksh showing 4x better performance on 1000 elements.