testlib: add test_suite class
Code is currently quite dirty, something that tremendously help
simplifying it would be an optional method in clash that if implemented
would get called each time an attribute is set. Another option would be
to generate for each attribute an is_set method, that would mean a lot
more generated code, but on the other hand this could be quite useful.