clash: rename and force constructor to be declared
Constructor now needs to be added to the list of methods and must be named
__init__ (and thus function implementation must be named <class name>__init__),
inspired on what python does.

Previous way to define a constructor was:

  class Foo  \
    attr1    \
    attr2    \
    _method1 \
    _method2 \

  _Foo() {
    # constructor implementation here...
  }

  Foo_method1() {
  ...

What was nice here was that the constructor didn't need to be part of the
actual class declaration, it was automatically attempted to be called, and that
was the actual reason why this needed to be changed.

Now:

  class Foo  \
    attr1    \
    attr2    \
    _method1 \
    _method2 \
    __init__ \ <---  needs to be added to the class declared methods

  Foo__init__() {
    # constructor implementation here...
  }

There was no way to detect if the constructor had been defined or not without
executing it, that meant that stderr was suppressed to avoid the "cmd not
found" message if the constructor was omitted, but also meant that if the
constructor was implemented then any stderr output would also be suppressed.

The exact inner call from clash when creating an object was:

  _'"$name"' "$@" 2> /dev/null || :

Where the "name" variable contained the class name.

One other way to fix this would have been to define a dummy one on class
creation, but that would override any previously defined constructor, which was
not acceptable.