clash: compat: replace $local by a generated alias
Previous style:

  "$local" var=...

Now replaced by:

  local var=...

Where local is an alias to a compatible builtin when the shell does not support
it (ksh/mksh).

For shells that support local, the actual local is directly used, no alias is
created.

It hasn't been tested, but in theory shells that do not support any of
local/typeset/declare should be compatible, local would be an alias to
nothing, and would lead to a simple global assignment.

Some changes in the source order have been done, notably clash should be
sourced at the top level of the program, not inside another lib (like
from withing lib/list like it was done before). This is due to ksh
aliases not getting expanded withing a source.

The code to test whether local is compatible and that declares aliases
is not testing any shell in particular but testing instead if features
are available. This is more tedious to do but is obviously more
reliable. This was advised by Andy Chu during a discussion on osh.