clash: osh full support
This was mostly thanks to commit 52c8ca7 ("$local" replacement by an
alias).
The last two tweaks in this commit are:
  local "$var"=...
osh does not allow dynamic assignment, the var name must be constant.
  $((a + 1 ))
`a' cannot be unset, osh does not cast those values to 0 if they are
either unset or NAN.