clash: fix ksh getattr infinite recursion
functions defined with the `function` keyword in ksh have typeset
defined variables local to their scope only. Meaning:

  function f {
    typeset var=foo
    g
  }

  function g {
    echo "var: $var"
  }

  f # output -> var:

so here var is local to f, and won't be inherited to g. To do so g would
need to be defined with the POSIX style g() { ... } .

So issue is that we need to define ksh functions with the `function`
keyword to have local variables (otherwise they are global) and that __getattr__
being a simple method means that it will call the getters to get its context,
which means each getter will call getattr etc. tl;dr infinite recursion.

To prevent that we set an internal var `__allow_getattr` to false in the
methods context generation to prevent getters from calling getattr in
this specific case (it makes sense since these are internal calls, the
user should have no knowledge that this is how methods get their variables
declared, and thus shouldn't expect getattr to be invoked).

But in ksh case getattr is defined with `function` and has recursion,
which means a `function` defined function calling another `function`
defined function which means that local variables are not inherited and
thus the allow_getattr value not passed.
So basically ksh cannot do recursion with `function` defined functions
without making use of function args or using a global var, relying on
inheriting local vars will not work.

The fix is simply to make this var global, as explained in the code
comment we set and unset this var and only call the getters which only
do printf calls, so there should be no frame unexpectedly modifying the var of
another frame, so while this is still a dirty hack this should be relatively
safe.