lib: list: add remove_node_ for direct removal
This is useful for dict, which already uses append_node_ to be able to
insert and remove a node by just holding a reference to it, and most
importantly without needing to iterate over the bucket list.
The basic idea is that the node is named with the hash, so we can
retrieve the node by just calculating the hash, then we can just remove
it or append another one without having to iterate the list it belongs
to.