lib: dict: cleanup empty buckets on items removal
This also contains a fix for remove_node_ call on a list with a single
element.

  x - [ 1, 2 ]
  |
  y - [ 3 ]
  |
  z - [ 4, 5 ]

When removing 3 we were previously just doing:

  x - [ 1, 2 ]
  |
  y - []
  |
  z - [ 4, 5 ]

That would potentially lead to a bunch of empty buckets that do not hold
any values.

Now that List_remove_node_ is available we can check if the bucket is
empty when removing an item, if that's the case we can remove the bucket
altogether. This is done without iterating over the buckets.

  x - [ 1, 2 ]
  |
  z - [ 4, 5 ]