lib: list: add sort method
Merge sort algorithm, no new nodes added, new list views are created
with a modified length, head and tail, so we always operate on the
actual list, no data is copied.

A new List_view class was introduced for the merge sort, since
we divide the list in two until we get a single element the number of
List objects created can be very high for big lists. This means
instantiating for each object a set of _map _pop etc. methods which are
unused in the sort context. This heavily slows down the shell and adds a
lot of memory consumption. The workaround is that List_view class which
only has a sort method and the needed head/tail/length attributes.
This helped improve performance tremendously, escpecially for shells
that don't scale well such as ksh (20x faster for 1000 elts).

The current sort only supports integer values, I would need to add
something similar to __lt__ support like in python to be able to sort
anything. Side note: string sorting would be just a matter of changing
the operator to lexicographic compare.