cleanup: function check with `command -v'
Previously thought the only way to check if a function existed was
either to use the non POSIX declare/typeset -f, or to simply run it and
check for command not found return value (127). Turns out that command
-v can do it, this is much simpler than redirecting "command not found"
to /dev/null and modifying the PATH to prevent multiple execve bound to
fail (and this `command' behavior is specified by POSIX).