test: core: handle ksh 2020 detection
core tests weren't passing due to ksh not skipping frame leak checks
anymore. This was due to a ksh version change between fedora 30 and 32,
the 2020 version has no mention of ksh93 anymore, so let's just rely on
check if $KSH_VERSION is not empty and not mksh.

For reference $KSH_VERSION value is: Version A 2020.0.0