clash: change constructor design
This is a big change to clash overall design:

Previous constructor behavior:
------------------------------

Whether __init__ is present in the class declared methods or not,
attributes will receive positional arguments values in a sequential
order, but attributes ending with an underscore will be ignored since we
do not always want to assign constructor args.

Example:

  class Car  \
    speed    \
    nb_used_ \
    brand    \

  Car 200 Toyota

Here nb_used_ is ignored due to the trailing underscore, since it did
not make sense to assign it a value when creating the object.
speed is 200 and brand is Toyota.

Adding __init__ to do a custom constructor wouldn't change much,
attributes are still assigned their values based on positional arguments
even before __init__ is called.

The drawback of this construction behavior is that it forces to add a
trailing underscores to many attributes, for instance List has length_,
which is quite inconvenient.
One other solution would be to assign them a new value in __init__
but that is both dirty and error prone, for instance:

  class List \
    length   \
    [...]
    __init__ \

  List__init__() {
    # length is initially equal to the first list item
    "$self"_length_is "$#"
  }

  List letters a b c d e f [...]

New constructor behavior:
-------------------------

When using a custom __init__, i.e. when adding __init__ as part of the
class declaration, then no assignment is done to any argument,
everything must be done by the user.

Now that we have a way to avoid automatic assignment we can get rid of
the underscore rule, trailing underscore now doesn't have any
special meaning.

So when __init__ is not used, all attributes receive their values
sequentially with the positional arguments, trailing underscore or not,
no argument is skipped (except methods, which start with an underscore).
Note that there is still no underlying __init__ declared in this case,
if __init__ is not in the declaration list then calling obj__init__ is
not possible, this is to avoid potentially overriding a user defined
__init__ method before the actual corresponding class creation, this
works for other methods so we should avoid breaking this for __init__ as
well.

--------------
- examples and README have been updated accordingly, now assigning
values in their constructor
- core tests have been updated as well (quite a big change)
- unrelated small changes:
  - renamed quote function to __clash_quote
  - added a few more comments to clash file