clash: hybrid mode: use variables to reduce cmd substitutions
This is an attempt to use global variables in the underlying structure
of clash to avoid sequential command substitutions when preparing a
method's frame. The idea came when writing about clash performance in
the Q/A section of the readme.

As seen in the example below the hybrid mode is much much simpler, and
avoids the dreaded getter calls to populate the local variables.

Though from a few tests it doesn't seem to be improving performance by a
big margin (if at all), and the number of forks only went down by 1/4 ~
1/5, which is less than I initially expected.

Classic mode:

  my_car_speed() {
    printf %s '200'
  }

  my_car_speed_is() {
    if [ "$self" = my_car ]; then
      speed=$1
    fi

    local _tmp_getter="$1"
    case "$1" in
      *"'"*)
        _tmp_getter="$(__clash_quote "$1"; printf x)";
        _tmp_getter="${_tmp_getter%x}"
        ;;
    esac

    eval my_car_speed'() {
      printf %s '"'$_tmp_getter'"'
    }'
  }

  my_car_start()
  {
    if [ -z "$_upper_object_id" ] || [ "$_upper_object_id" != my_car ]; then
      local speed="$(my_car_speed; printf x)"
      speed="${speed%x}"
      local brand="$(my_car_brand; printf x)"
      brand="${brand%x}"
      local self=my_car
      local _upper_object_id=my_car
    fi
    Car_start "$@"
  }

Hybrid mode:

  my_car_speed() {
      printf %s "$__clash_my_car_speed"
  }

  my_car_speed_is() {
      if [ "$self" = my_car ]; then
          speed=$1
      fi
      __clash_my_car_speed=$1
  }

  my_car_start() {
      if [ -z "$_upper_object_id" ] || [ "$_upper_object_id" != my_car ]; then
          local speed="$__clash_my_car_speed"
          local brand="$__clash_my_car_brand"
          local self=my_car
          local _upper_object_id=my_car
      fi
      Car_start "$@"
  }