clash: support upper frame leaking for ksh93
Details in the code comment of this commit:

```
For upper frame leaking variable to work on ksh93 we specifically need
to have the caller (a method) defined with the `function foo {}` style,
and the callee (a setter) to have its method defined with the POSIX
style. Otherwise with both of them as `function ...{}` the variable will
either leak in the global scope if typeset isn't used, or just stay
local if typeset is used. This is the reason why this function
declaration is the only one hardcoded with the POSIX style.
```

Thinking about it again, with ksh93, if the caller of the setter is also
defined with the POSIX style then that means that calling a setter will
set the variable in the global scope, all classes in lib are currently
defined that way. Though if they are defined as such then that means
that their own typeset is not generating any local var, so this is
already a pretty bad situation.