lib: remove trailing underscores for "public" attrs
This is a follow-up on the constructor design change (e3957a8) which
removed the trailing underscore special meaning.

Since it was conflicting with the underscore removal, also remove the
`hash_function` function that was used to verify the dict was empty
before allowing to change the hash function. Either way it felt weird to
use a method to set an attribute, if needed to value check I might look
into using the __setattr__ hook, but I'm not sure if the current
implementation would allow me to bail out if I found a constraint
violation, maybe I should revisit the way it works.