lib: move set to examples
Considering how slow the current implementation is might as well move it
to the examples for now.

This performance issue is mostly due to clash generating an insane
amount of code when dealing with big recursive data structures (the
RedBlackTree class here), putting the shell on its knees.

This is the exact same reason that lead to the creation of a stripped
down List class called ListView to be able to handle List merge sort,
which showed 3x speed improvements.